% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipes.R
\docType{data}
\name{pipes}
\alias{pipes}
\title{Pipe sequence data}
\format{a list per source file of each magrittr chain found, with character
vector elements of the function names in the chain.}
\usage{
pipes
}
\description{
The function sequence of all of the packages using magrittr pipes.
}
\keyword{datasets}

